+++ hblanc/move.cpp
@@ -1,0 +1,24 @@
+#include"move.hpp"
+
+int move(int dir)
+{
+    const Uint8 *keystates = SDL_GetKeyboardState(NULL);
+    int new_dir = dir;
+    if (keystates[SDL_SCANCODE_UP])
+    {
+       new_dir = UP;
+    }
+    if (keystates[SDL_SCANCODE_DOWN])
+    {
+        new_dir = DOWN;
+    }
+    if (keystates[SDL_SCANCODE_LEFT])
+    {
+        new_dir = LEFT;
+    }
+    if (keystates[SDL_SCANCODE_RIGHT])
+    {
+       new_dir = RIGHT;
+    }
+    return (dir == -new_dir) ? dir : new_dir;
+}
