+++ hblanc/move.hpp
@@ -1,0 +1,9 @@
+#pragma once
+#define MOVECASE 10
+
+int snakeY = 400;
+int snakeX = 400;
+#define UP -1
+#define DOWN 1
+#define RIGHT -2
+#define LEFT 2
