+++ hblanc/segment.hpp
@@ -1,6 +1,21 @@
 #pragma once
+#define MOVECASE 10
 
-class Segment {
+class Segment
+{
+public:
+  Segment(int x, int y);
+  ~Segment();
+
+  int GetX();
+  int GetY();
+  void SetX(int);
+  void SetY(int);
+
+  Segment *GetNext();
+  void SetNext(Segment *next);
+
 private:
-  Segment* next;
+  int x, y;
+  Segment *next;
 };
